move tests to allinon
