+++ projects/batfish/src/test/java/org/batfish/grammar/cisco/CiscoGrammarTest.java
@@ -5326,63 +5326,4 @@ public class CiscoGrammarTest {
     assertThat(ccae, hasUndefinedReference(filename, NETWORK_OBJECT, "undef-source-mapped"));
     assertThat(ccae, hasUndefinedReference(filename, NETWORK_OBJECT, "undef-source-real"));
   }
-
-  @Ignore
-  @Test
-  public void testNatTable() throws IOException {
-    String hostname = "ios-nat-table";
-
-    Configuration configuration = parseConfig(hostname);
-    Interface inside = configuration.getAllInterfaces().get("FastEthernet0/0");
-    Interface outside = configuration.getAllInterfaces().get("GigabitEthernet1/0");
-
-    Flow flow =
-        Flow.builder()
-            .setSrcIp(Ip.parse("1.0.0.2"))
-            .setSrcPort(10000)
-            .setDstIp(Ip.parse("2.0.0.1"))
-            .setDstPort(20000)
-            .setIngressNode(hostname)
-            .setIngressInterface(inside.getName())
-            .setIpProtocol(IpProtocol.TCP)
-            .setTag("tag")
-            .build();
-
-    Batfish batfish = getBatfishForConfigurationNames(hostname);
-    batfish.computeDataPlane();
-    Map<Flow, List<TraceAndReverseFlow>> traces =
-        batfish.getTracerouteEngine().computeTracesAndReverseFlows(ImmutableSet.of(flow), false);
-
-    assertThat(traces.get(flow), hasSize(1));
-
-    Flow returnFlow = traces.get(flow).get(0).getReverseFlow();
-    assertThat(
-        returnFlow,
-        equalTo(
-            Flow.builder()
-                .setDstIp(Ip.parse("10.0.0.1"))
-                .setDstPort(10000)
-                .setSrcIp(Ip.parse("2.0.0.1"))
-                .setSrcPort(20000)
-                .setIngressNode(hostname)
-                .setIngressInterface(outside.getName())
-                .setIpProtocol(IpProtocol.TCP)
-                .setTag("tag")
-                .build()));
-
-    Set<FirewallSessionTraceInfo> sessions = traces.get(flow).get(0).getNewFirewallSessions();
-
-    Map<Flow, List<TraceAndReverseFlow>> returnFlowTraces =
-        batfish
-            .getTracerouteEngine()
-            .computeTracesAndReverseFlows(ImmutableSet.of(returnFlow), sessions, false);
-
-    assertThat(returnFlowTraces.get(returnFlow), hasSize(1));
-    List<TraceAndReverseFlow> returnTraceAndReverseFlows = returnFlowTraces.get(returnFlow);
-
-    // return flow should be null routed given the null route in the config
-    assertThat(
-        returnTraceAndReverseFlows.get(0).getTrace().getDisposition(),
-        equalTo(FlowDisposition.NULL_ROUTED));
-  }
 }
